+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -28,5 +28,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("8.5.0.5")]
-[assembly: AssemblyFileVersion("8.5.0.5")]
+[assembly: AssemblyVersion("8.5.0.6")]
+[assembly: AssemblyFileVersion("8.5.0.6")]